move test_scalar_wrapper to test_buffer.py
